refactor: split dependency stack helpers
